Capture and return errors from all function calls
Prior to this we ignored quite a lot of errors by assigning them to
the `_` identifier.  Unfortunately this causes crashes when things go
wrong, eg https://github.com/ncw/rclone/issues/3140

Using `errcheck -blank` all these places were found and the errors
propagated to the caller, changing the interfaces as necessary.

Some places were not changed, eg `resp.Body.Close()` as they are very
unlikely to return an error and if they do it will not have
consequences.