Fix base64 decoding errors: "illegal base64 data at input byte 22"
Sometimes mega seems to use Standard encoding instead of URL encoding
for base64 and their C++ decoder accepts both so this change
implements that.

This also changes the interface so that base64 input is passed as
string rather than []byte which saves a lot of casting to []byte and
back.

See: https://forum.rclone.org/t/problem-to-login-with-mega/12276